[PM-32688] Re-organize conversions.rs
